Fixed paths for config and keys files
