+++ src/classes/bullet.js
@@ -1,5 +1,5 @@
 import { GameObjects } from 'phaser';
-import ASSETS_KEYS from '../components/assets-keys';
+import ASSETS_KEYS from '../keys/assets-keys';
 
 export default class Bullet extends GameObjects.Sprite {
   constructor(scene) {
