--- src/components/assets-keys.js
@@ -1,16 +1,0 @@
-const ASSETS_KEYS = {
-  SHIP: 'ship',
-  BULLET: 'bullet',
-  SPACE_BACKGROUND: 'spaceBackground',
-  LOGO: 'phaserLogo',
-  BUTTON: 'button',
-  BUTTON_ON_HOVER: 'buttonOnHover',
-  UNCHECKED_BOX: 'uncheckedBox',
-  CHECKED_BOX: 'checkedBox',
-  ASTEROID: 'asteroid',
-  EXPLOSION: 'explosion',
-  BG_MUSIC: 'bgMusic',
-  PIXEL_FONT: 'pixelFont',
-};
-
-export default ASSETS_KEYS;
