--- src/components/scene-keys.js
@@ -1,10 +1,0 @@
-const SCENE_KEYS = {
-  GAME: 'Game',
-  CREDITS: 'Credits',
-  BOOT: 'boot',
-  PRELOADER: 'Preloader',
-  TITLE: 'Title',
-  OPTIONS: 'Options',
-};
-
-export default SCENE_KEYS;
