+++ src/scenes/title-scene.js
@@ -1,8 +1,8 @@
 /* eslint-disable class-methods-use-this */
 import { Scene, Display } from 'phaser';
-import CONFIG from '../components/config';
-import SCENE_KEYS from '../components/scene-keys';
-import ASSETS_KEYS from '../components/assets-keys';
+import CONFIG from '../config';
+import SCENE_KEYS from '../keys/scene-keys';
+import ASSETS_KEYS from '../keys/assets-keys';
 
 export default class TitleScene extends Scene {
   addBackground() {
